+++ example/lib/src/view.dart
@@ -1,0 +1,5 @@
+// Copyright 2021 Andrious Solutions Ltd. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+export 'package:example/src/menu/app_menu.dart';