+++ CHANGELOG.md
@@ -1,9 +1,19 @@
-## 1.0.0
- December 14, 2021
-- initial commit and release
+
+## 1.2.0
+ December 21, 2021
+- Example app
+- class WebPage to class WebPageWidget
+- PreferredSizeWidget? appBar, in WebPageController
+- PreferredSizeWidget? onAppBar() =>
+- final bool? primary;
 
 ## 1.1.0
  December 17, 2021
 - utils/hyperlink.dart';
 - utils/iframe_widget.dart';
 - utils/web_utils.dart';
+
+## 1.0.0
+ December 14, 2021
+- initial commit and release
+